+++ packages/Flow/src/package-test.ts
@@ -18,6 +18,7 @@ import './tests/guide-tests';
 import './tests/summary-tests';
 import './tests/files-tree-tests';
 import './tests/execution-preview-tests';
+import './tests/viewer-tests';
 
 export let _package = new DG.Package();
 export { tests };